Update Solar app to Corticon.js V1.1.
Remove the unneeded custom index.js files + updated the one for savings: With Corticon V1.0 we didn't pass through data that did not belong to rules. So we had to customize  all the index.js files.  However as of Corticon V1.1, all data is passed through, so the only custom step left is for the solar savings rulesheet (as the entry point for that rulesheet needs to create the payload from the output of a parallel step).

Updated all assets from studio.
Added a zip file with the assets for each version (will be easier to access than restoring from history).