Add the new decision service bndle corresponding to previous changes:
Add a vocabulary attribute to show time as art of the DateTime UI control.
Also add an attribute to allow sorting options - this will be usefull for data we get from RESt services as it may come in random order.

Update Step3 to contain a DateTime control that shows the time part.  And add various combinations of DateTime control with min or max or both or none.

The client side component has not been updated yet to handle all of these.  Watch for a future commit.