Add JSON extraction and insertion SCO example
This commit introduces a new sample demonstrating how to extract and insert entire datasets as JSON in the in-memory payload. It includes two service callouts: one for adding entities and their associations to the payload, and another for retrieving this data in JSON form. This example is backend agnostic, allowing it to run without dependencies and be tested in Corticon Studio.